requirements.txt
numpy>=1.23
Pillow>=9.5
cryptography>=41.0
opencv-python>=4.7
